+++ modules/AudioModule.cpp
@@ -22,7 +22,7 @@
 
 #include "AudioModule.h"
 
-AudioModule::AudioModule(qint64 samplerate, ModuleControls *controls ) :
+AudioModule::AudioModule(int samplerate, ModuleControls *controls ) :
 	AudioProcess( samplerate ),
 	m_controls( controls )
 {
@@ -34,7 +34,7 @@ AudioModule::~AudioModule()
 
 }
 
-void AudioModule::processAudio(sampleFrame *buffer, qint64 len)
+void AudioModule::processAudio(sampleFrame *buffer, int len)
 {
 	memset( buffer, 0, sizeof(sampleFrame) * len );
